finish

+++ Binomial.java
@@ -20,7 +20,7 @@ public class Binomial {
 	
 	// Computes the Binomial function, efficiently
 	public static long binomial(int n, int k) {
-		int BIN [][] = new int[n+1][k+1];
+		long BIN [][] = new long[n+1][k+1];
 		for (int i=0; i<BIN.length; i++){
 			for (int j=0; j< BIN[0].length; j++){
 				BIN[i][j] = -1;
@@ -33,7 +33,7 @@ public class Binomial {
 		//// Replace the following statement with your code.
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
